Добавит компонент карточка предложения
