+++ src/components/app/app.tsx
@@ -5,22 +5,21 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import OfferPage from '../pages/offer-page';
 import NotFound from '../pages/not-found';
 import PrivateRoute from '../private-route';
+import { Offer } from '../../types/offer';
 
 type AppProps = {
   proposalsCount: number;
+  offers: Offer[];
 };
 
-export default function App({
-  proposalsCount: proposalsCount,
-}: AppProps): JSX.Element {
+export default function App({ proposalsCount, offers }: AppProps): JSX.Element {
   return (
     <BrowserRouter>
       <Routes>
         <Route
           path="/"
-          element={<MainPage proposalsCount={proposalsCount} />}
-        >
-        </Route>
+          element={<MainPage proposalsCount={proposalsCount} offers={offers} />}
+        />
         <Route path="/offer/:id" element={<OfferPage />}></Route>
         <Route path="/login" element={<LoginPage />}></Route>
         <Route
@@ -30,8 +29,7 @@ export default function App({
               <FavoritesPage />
             </PrivateRoute>
           }
-        >
-        </Route>
+        />
         <Route path="*" element={<NotFound />}></Route>
       </Routes>
     </BrowserRouter>
