+++ src/components/offer-card.tsx
@@ -1,0 +1,65 @@
+import { Offer } from '../types/offer';
+
+type OfferCardProps = {
+  offer: Offer;
+};
+
+type RaitingProps = {
+  rating: number;
+}
+
+function Rating({rating}: RaitingProps): JSX.Element {
+  const ratingCeil = Math.round(rating);
+  const width = ratingCeil * 20;
+  return (
+    <div className="place-card__rating rating">
+      <div className="place-card__stars rating__stars">
+        <span style={{ width: `${width}%` }} />
+        <span className="visually-hidden">Rating</span>
+      </div>
+    </div>
+  );
+}
+
+export default function OfferCard({ offer }: OfferCardProps): JSX.Element {
+  const { isPremium, previewImage, price, title, type, rating } = offer;
+  return (
+    <article className="cities__card place-card">
+      {isPremium && (
+        <div className="place-card__mark">
+          <span>Premium</span>
+        </div>
+      )}
+      <div className="cities__image-wrapper place-card__image-wrapper">
+        <a href="#">
+          <img
+            className="place-card__image"
+            src={previewImage}
+            width={260}
+            height={200}
+            alt="Place image"
+          />
+        </a>
+      </div>
+      <div className="place-card__info">
+        <div className="place-card__price-wrapper">
+          <div className="place-card__price">
+            <b className="place-card__price-value">€{price}</b>
+            <span className="place-card__price-text">/&nbsp;night</span>
+          </div>
+          <button className="place-card__bookmark-button button" type="button">
+            <svg className="place-card__bookmark-icon" width={18} height={19}>
+              <use xlinkHref="#icon-bookmark" />
+            </svg>
+            <span className="visually-hidden">To bookmarks</span>
+          </button>
+        </div>
+        <Rating rating={rating} />
+        <h2 className="place-card__name">
+          <a href="#">{title}</a>
+        </h2>
+        <p className="place-card__type">{type}</p>
+      </div>
+    </article>
+  );
+}
