+++ src/index.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/app/app';
 import { Settings } from './const';
+import { offers } from './mocks/offer';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -9,6 +10,6 @@ const root = ReactDOM.createRoot(
 
 root.render(
   <React.StrictMode>
-    <App proposalsCount={Settings.ProposalsCount} />
+    <App proposalsCount={Settings.ProposalsCount} offers={offers}/>
   </React.StrictMode>
 );
