+++ src/mocks/offer.ts
@@ -1,25 +1,4 @@
-type City = {
-  name: string;
-  location: Location;
-};
-type Location = {
-  latitude: number;
-  longitude: number;
-  zoom: number;
-};
-
-type Offer = {
-  id: string;
-  title: string;
-  type: string;
-  price: number;
-  previewImage: string;
-  city: City;
-  location: Location;
-  isFavorite: boolean;
-  isPremium: boolean;
-  rating: number;
-};
+import { Offer } from '../types/offer';
 
 export const offers: Offer[] = [
   {
